接口实现

+++ service/src/main/java/com/huotu/loanmarket/service/Repository/LoanUserRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by hxh on 2017-10-24.
  */
-public interface LoanUserRepository extends JpaRepository<LoanUser,Integer>{
+public interface LoanUserRepository extends JpaRepository<LoanUser,Long>{
 }
